requirements.txt
annotated-types==0.6.0
anyio==4.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
distro==1.8.0
Flask==3.0.0
fpdf==1.7.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
openai==1.3.8
pandas==2.1.1
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
wheel==0.41.2
